feat(react): refactor preact upstream tests into compiled/no-compile split
- Split setup.js into setup-shared.js, setup-nocompile.js, setup-compiled.js
  for cleaner separation between the two test modes
- Update skiplist.json with refined categorization across both modes
- Update vitest.shared.ts with improved transform plugin
- Update vitest.compiled.config.ts for the new setup file names
- Update README with comprehensive docs on both modes and skiplist workflow
- Add GitHub Copilot instructions for skiplist maintenance
